Add files via upload
This is the data used to create the figures and charts of the article titled "Machine Learning Assisted Industrial Symbiosis: Testing the Ability of Word Vectors to Estimate Similarity for Material Substitutions".  The corpus data that support the findings of this study are available from AcclaimIP (patents corpus) and the Elsevier's Text Mining API (scientific articles). Restrictions apply to the availability of these data, which were used under license for this study. Data are available at https://www.acclaimip.com/ and https://dev.elsevier.com/ respectively with the access granted from Acclaim and Elsevier.